integration-tests/requirements.txt
behave==1.*
black
environs
faker==4.*
isort
kombu==4.*
mypy
python-jose==3.*
PyYAML==5.3.1
reportportal-behave-client==1.*
requests==2.*
she-data-generation>=1.0.5
she-logging==1.*
